+++ src/main/java/lotto/model/WinningNumbers.java
@@ -1,8 +1,8 @@
 package lotto.model;
 
-import static lotto.model.LottoRule.MAX_NUMBER;
-import static lotto.model.LottoRule.MIN_NUMBER;
-import static lotto.model.LottoRule.NUMBER_COUNT;
+import static lotto.model.constant.LottoRule.MAX_NUMBER;
+import static lotto.model.constant.LottoRule.MIN_NUMBER;
+import static lotto.model.constant.LottoRule.NUMBER_COUNT;
 
 import java.util.List;
 import lotto.dto.WinningNumberRequestDTO;
